pix2pix_project/pix2pix_project/requirements.txt
tensorflow>=2.10.0
opencv-python>=4.7.0
numpy>=1.23.0
flask>=2.3.0
Pillow>=9.5.0
matplotlib>=3.7.0
tqdm>=4.65.0
seaborn>=0.12.0
